Reorganize the directories and update _vimrc
Organized the directories so that vim would automatically load the
items. Created the colors and plugin directory to do this.

Updated the Windows _vimrc so that it would use pathogen to update the
runtimepath and load everything appropriately.

I'll update the *nix .vimrc from another machine to use pathogen and
upload it at a later date.